+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 most recent version is listed first.
 
 
+## **version:** v0.7.4
+- Fix a bug that prevented `enquire_link` and `deliver_sm` from been sent: https://github.com/komuw/naz/pull/179  
+- Enforce `naz` message protocol in code : https://github.com/komuw/naz/pull/179
+
+
 ## **version:** v0.7.3
 - make `naz.SimpleCorrelater.delete_after_ttl()` private: https://github.com/komuw/naz/pull/172
 - rename `naz.broker.BaseOutboundQueue` to `naz.broker.BaseBroker`: https://github.com/komuw/naz/pull/174